refactor(plugin): move shared ai proxy handler to shared module
